requirements.txt
behavex==4.0.9
python-dotenv==0.19.2
pandas==2.2.2